requirements.txt
absl-py==1.0.0
access==1.1.7
affine==2.3.1
alembic==1.7.7
altair==4.2.0
amply==0.1.5
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
bokeh==2.4.3
Bottleneck==1.3.4
branca==0.5.0
brewer2mpl==1.4.1
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
Cartopy==0.20.0
certifi==2022.5.18.1
certipy==0.1.3
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.1.0
colorama==0.4.4
conda==4.12.0
conda-package-handling==1.8.1
cryptography==37.0.2
cupy==10.4.0
cycler==0.11.0
Cython==0.29.30
cytoolz==0.11.2
dask==2022.5.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
dill==0.3.5.1
distributed==2022.5.2
docutils==0.18.1
entrypoints==0.4
esda==2.4.1
executing==0.8.3
fastjsonschema==2.15.3
fastrlock==0.8
Fiona==1.8.20
flatbuffers==2.0
flit_core==3.7.1
folium==0.12.1.post1
fonttools==4.33.3
fsspec==2022.5.0
gast==0.5.3
GDAL==3.3.2
geopandas==0.10.2
giddy==2.3.3
gmpy2==2.1.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.46.3
h5py==3.6.0
HeapDict==1.0.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.19.2
imbalanced-learn==0.9.1
importlib-metadata==4.11.4
importlib-resources==5.7.1
inequality==1.0.0
ipykernel==6.13.0
ipyleaflet==0.16.0
ipympl==0.9.1
ipyparallel==8.3.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsmin==3.0.1
json5==0.9.5
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.17.0
jupyter-telemetry==0.1.0
jupyterhub==2.3.0
jupyterlab==3.4.2
jupyterlab-drawio==0.9.0
jupyterlab-logout==0.5.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-topbar==0.6.1
jupyterlab-widgets==1.1.0
keplergl==0.3.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
laspy==2.1.2
libclang==14.0.1
libmambapy==0.23.3
libpysal==4.6.2
llvmlite==0.36.0
locket==1.0.0
lxml==4.8.0
lz4==4.0.0
Mako==1.2.0
mamba==0.23.3
mapclassify==2.4.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
metakernel==0.29.0
mgwr==2.1.2
mistune==0.8.4
momepy==0.5.3
mpmath==1.2.1
msgpack==1.0.3
munch==2.5.0
munkres==1.1.4
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nbresuse==0.4.0
nest-asyncio==1.5.5
networkx==2.8.2
nose==1.3.7
notebook==6.4.11
notebook-shim==0.1.0
numba==0.53.1
numexpr==2.7.3
numpy==1.21.6
oauthlib==3.2.0
octave-kernel==0.34.2
opt-einsum==3.3.0
osmnx==1.2.0
packaging==21.3
pamela==1.0.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pip==22.1.1
pkgconfig==1.5.5
plotly==5.8.0
pointpats==2.2.0
portalocker==2.4.0
postgres-kernel==0.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pyclustering==0.10.1.2
pycosat==0.6.3
pycparser==2.21
pycurl==7.45.1
pygeos==0.10.2
Pygments==2.12.0
PyGObject==3.42.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
pysal==2.6.0
pyshp==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.0.0
quantecon==0.5.2
rasterio==1.2.8
rasterstats==0.16.0
requests==2.27.1
requests-oauthlib==1.3.1
rise==5.7.1
rpy2==3.5.1
rsa==4.8
Rtree==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.80
rvlib==0.0.6
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
segregation==2.1.0
Send2Trash==1.8.0
setuptools==62.3.2
shapely==1.8.2
simplegeneric==0.8.1
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.1
spaghetti==1.6.5
spglm==1.0.8
spint==1.0.7
splot==1.1.5.post1
spopt==0.3.0
spreg==1.2.4
spvcm==0.3.0
SQLAlchemy==1.4.36
stack-data==0.2.0
statsmodels==0.13.2
sympy==1.10.1
tables==3.7.0
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tobler==0.9.0
toolz==0.11.2
torch==1.11.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torch-spline-conv==1.2.1
torchaudio==0.11.0
torchvision==0.12.0
torchviz==0.0.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
traittypes==0.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
unicodedata2==14.0.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.0
wrapt==1.14.1
xlrd==2.0.1
xyzservices==2022.4.0
zict==2.2.0
zipp==3.8.0
zstandard==0.17.0
